Move add/invite member button to page actions
